tests/wpt/tests/html/semantics/embedded-content/media-elements/offsets-into-the-media-resource/duration.html

<!doctype html>
<title>duration</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  var v = document.createElement('video');
  assert_true(isNaN(v.duration));
}, 'duration initial value');
</script>
